requirements.txt
PyExifTool>=0.5.6
reportlab
pypandoc
pyyaml>=6.0
